Added catch to make sure program does not crash when loading wrong data file in Rietveld and started to split graphicsviewlib
